moved RCAttributionNetwork declarations to separate header
